csv-extractor-simple/pyproject.toml
[tool.poetry]
name = "csv-extractor-simple"
version = "1.0.0"
description = "A simple CSV extractor"
authors = ["Mathias Lohne <mathias.lohne@cognite.com>", "Marie Solvik Lepoutre <marie.lepoutre@cognite.com>"]

[tool.poetry.dependencies]
python = "^3.7"
cognite-extractor-utils = "^2.2.0"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
flake8 = "^3.9.2"
mypy = "^0.910"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
csv_extractor = "csv_extractor.extractor:main"
